Generate better error messages on invalid options
When attempting to set an option to an invalid value, libssh always
generates the message “Invalid argument in ssh_options_set”. This is not
really useful in contexts where the stack trace is not available.

libssh-ruby will now generate a more explanative error message when
ssh_options_set returns < 0, by including the option we were trying to
set along with the invalid value.

The raised exception will be an ArgumentError rather than a
LibSSH::Error, because that’s what it is.